+++ generators/control/index.js
@@ -17,6 +17,7 @@ const {
 const controlDependencies = require( './dependencies.json' );
 const technologies = require( '../../lib/technologies' );
 const commonPrompts = require( '../../lib/common-prompts' );
+const { LICENSE_UNLICENSED } = require( '../../lib/constants' );
 
 module.exports = class extends Generator {
 
@@ -50,6 +51,7 @@ module.exports = class extends Generator {
          angularDirectiveName: '',
          description: '',
          license: this.config.get( 'license' ),
+         privateModule: '',
          homepage: this.config.get( 'homepage' ),
          author: this.config.get( 'author' ),
          integrationTechnology: this.config.get( 'laxarIntegrationTechnology' ) || 'plain',
@@ -89,6 +91,7 @@ module.exports = class extends Generator {
 
             this.vars.cssClassName = this.vars.name.replace( /[_\s]+/, '-' );
             this.vars.banner = createBanner( this );
+            this.vars.privateModule = this.vars.license === LICENSE_UNLICENSED ? '\n  "private": true,' : '';
 
             this.technology =
                technologies.find( _ => _.integrationTechnology === this.vars.integrationTechnology );