+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Last Changes
 
+- [#23](https://github.com/LaxarJS/generator-laxarjs2/issues/23): license: Changed default from none to UNLICENSED
+
+
 ## v2.0.2
 
 - [#20](https://github.com/LaxarJS/generator-laxarjs2/issues/20): React template: use new laxar-react-adapter API